tests/wpt/tests/html/semantics/forms/the-form-element/form-elements-sameobject.html

<!doctype html>
<meta charset="utf-8">
<title>Testing [SameObject] on the 'elements' attribute on the 'form' element</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<form>
  <input>
</form>

<script>
test(function() {
  var form = document.querySelector('form');
  var elements = form.elements;
  assert_equals(elements, form.elements);
  form.appendChild(document.createElement('input'));
  assert_equals(elements, form.elements);
}, "[SameObject] should apply to 'elements' attr on <form>");
</script>
